% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fndwords.R
\name{find_words}
\alias{find_words}
\title{Odgadnij słowo w krzyżówce}
\usage{
find_words(write_a_word, dictionary, report = FALSE, to_low = TRUE)
}
\arguments{
\item{write_a_word}{\itemize{
\item Jako argument należy wpisać słowo, które chcemy odgadnąć. W miejsce liter, które nie są znane, należy wpisać kropki.
}}

\item{dictionary}{\itemize{
\item Jako argument należy podać obiekt, który jest wektorem, w którym znajdują się słowa.
}}

\item{report}{\itemize{
\item Wartość TRUE oznacza, że do pliku txt. zostanie zapisany wynik funkcji. Ponadto utworzony zostanie folder results, w którym umieszczane będą wszystkie wyniki.
}}

\item{to_low}{\itemize{
\item Wartość TRUE sprawia, że wszystkie wielkie litery w wektorze dictionary zostaną zamienione na małe. Ten argument przydaje się m.in. w przypadku słów niemieckich, ponieważ rzeczowniki w tym języku są rozpoczęte wielkimi literami. Niewykorzystanie tego argumentu może doprowadzić do otrzymania niepełnych wyników.
}}
}
\value{
wektor wydzielonych słów.
}
\description{
Funckja pozwala na wydzielenie słów z danego zbioru wyrazów.
}
\examples{
find_words("v..y", c("nice", "very", "small", "package", "bro"), report = TRUE)
find_words("....", c("nice", "very", "small", "package", "bro"), report = FALSE)
}
